Implement most GDT, IDT, and IST bits, and refactor into Kernel object
Implement most of the bits for setting up a GDT and IST, and preliminary
work for the IDT. Initialize the GDT and the IST, loading the IST into
TR and the GDT into GDTR (and switching to the new selectors). Refactor
much of the kernel stand-up code into a new Kernel object.