Remove GKARG dependency in klog and make a SimpleFb for early console
This genericizes the KernLogger so that it isn't statically tied to the
global GKARG, and at run-time a new logger can be attached to replace
the current one. Early in the boot process, the SimpleFb inherits the
static UnsafeFrameBuffer passed from the initial-stage kernel loader,
but as soon as we initialize a real gfx output driver, we want to switch
over to it for the logging output.